matched count

+++ cpp/net/data_comm.hpp
@@ -252,7 +252,7 @@ public:
 //        int index = sdispls[i] + j;
 //        DataTuple<DENT> tuple(dense_local->cols);
 //        tuple.col = sending_vec[j];
-        sendbuf[index].col = sending_vec[j];
+        sendbuf[index].col = static_cast<uint64_t>(sending_vec[j]);
       }
 
       for (int j = 0; j < receiving_vec.size(); j++) {
